requirements.txt
flask==2.2.3
werkzeug==2.2.3
psycopg2-binary==2.9.5


